# S4C_Admin/temp_swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Study4Clone API",
    "description": "IELTS Mock Test Backend API",
    "version": "v1"
  },
  "paths": {
...
    "/api/TestAttempts/user/{userId}": {
      "get": {
        "tags": [
          "TestAttempts"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "PageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "FromDate",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "ToDate",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "SortOrder",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TestAttemptListDtoPagedList"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestAttemptListDtoPagedList"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestAttemptListDtoPagedList"
                }
              }
            }
          }
        }
      }
    },
    "/api/TestAttempts/{id}/exam": {
      "get": {
        "tags": [
          "TestAttempts"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Exam"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Exam"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Exam"
                }
              }
            }
          }
        }
      }
    },
    "/api/TestAttempts/{id}": {
      "get": {
        "tags": [
          "TestAttempts"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TestAttemptDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestAttemptDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestAttemptDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/TestAttempts/{id}/result": {
      "get": {
        "tags": [
          "TestAttempts"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TestResultDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestResultDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestResultDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/TestAttempts/start": {
      "post": {
        "tags": [
          "TestAttempts"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartTestRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/StartTestRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/StartTestRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TestAttemptDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestAttemptDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestAttemptDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/TestAttempts/{id}/submit": {
      "post": {
        "tags": [
          "TestAttempts"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitTestRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitTestRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitTestRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TestResultDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestResultDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestResultDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/TestAttempts/{id}/abandon": {
      "post": {
        "tags": [
          "TestAttempts"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/TestAttempts/{id}/pause": {
      "post": {
        "tags": [
          "TestAttempts"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PauseTestRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PauseTestRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PauseTestRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AttemptStatus": {
        "enum": [
          "InProgress",
          "Pending",
          "Completed",
          "Abandoned"
        ],
        "type": "string"
      },
...
      "Exam": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "slug": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "thumbnailUrl": {
            "type": "string",
            "nullable": true
          },
          "type": {
            "$ref": "#/components/schemas/ExamType"
          },
          "status": {
            "$ref": "#/components/schemas/ExamStatus"
          },
          "duration": {
            "type": "integer",
            "format": "int32"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "skills": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExamSkill"
            },
            "nullable": true
          },
          "attempts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TestAttempt"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ExamSkill": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "examId": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "skill": {
            "$ref": "#/components/schemas/SkillType"
          },
          "orderIndex": {
            "type": "integer",
            "format": "int32"
          },
          "duration": {
            "type": "integer",
            "format": "int32"
          },
          "exam": {
            "$ref": "#/components/schemas/Exam"
          },
          "sections": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExamSection"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ExamStatus": {
        "enum": [
          "Draft",
          "Published",
          "Archived"
        ],
        "type": "string"
      },
      "ExamType": {
        "enum": [
          "IeltsAcademic",
          "IeltsGeneral",
          "Toeic",
          "Other"
        ],
        "type": "string"
      },
...
      "PauseTestRequest": {
        "type": "object",
        "properties": {
          "answers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SubmitAnswerRequest"
            },
            "nullable": true
          },
          "timeRemaining": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "StartTestRequest": {
        "type": "object",
        "properties": {
          "examId": {
            "type": "string",
            "format": "uuid"
          },
          "forceNew": {
            "type": "boolean"
          },
          "selectedSkills": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SkillType"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "SubmitTestRequest": {
        "type": "object",
        "properties": {
          "answers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SubmitAnswerRequest"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "TestAttemptDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "examId": {
            "type": "string",
            "format": "uuid"
          },
          "examTitle": {
            "type": "string",
            "nullable": true
          },
          "startedAt": {
            "type": "string",
            "format": "date-time"
          },
          "completedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "status": {
            "$ref": "#/components/schemas/AttemptStatus"
          },
          "overallScore": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "listeningScore": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "readingScore": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "writingScore": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "speakingScore": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "timeRemaining": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "selectedSkills": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "TestAttemptListDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "examId": {
            "type": "string",
            "format": "uuid"
          },
          "examTitle": {
            "type": "string",
            "nullable": true
          },
          "startedAt": {
            "type": "string",
            "format": "date-time"
          },
          "completedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "status": {
            "$ref": "#/components/schemas/AttemptStatus"
          },
          "overallScore": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "selectedSkills": {
            "type": "string",
            "nullable": true
          },
          "isFullTest": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "TestAttemptListDtoPagedList": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TestAttemptListDto"
            },
            "nullable": true
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "totalCount": {
            "type": "integer",
            "format": "int32"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          },
          "hasPreviousPage": {
            "type": "boolean",
            "readOnly": true
          },
          "hasNextPage": {
            "type": "boolean",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "TestResultDto": {
        "type": "object",
        "properties": {
          "attemptId": {
            "type": "string",
            "format": "uuid"
          },
          "examId": {
            "type": "string",
            "format": "uuid"
          },
          "examTitle": {
            "type": "string",
            "nullable": true
          },
          "overallBandScore": {
            "type": "number",
            "format": "double"
          },
          "totalCorrect": {
            "type": "integer",
            "format": "int32"
          },
          "totalQuestions": {
            "type": "integer",
            "format": "int32"
          },
          "percentageScore": {
            "type": "number",
            "format": "double"
          },
          "startedAt": {
            "type": "string",
            "format": "date-time"
          },
          "completedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "timeTaken": {
            "type": "string",
            "format": "date-span"
          },
          "skillResults": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SkillResultDto"
            },
            "nullable": true
          },
          "questionResults": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QuestionResultDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Enter your JWT token",
        "scheme": "Bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}